Refactor data processing and add new plotting function
The getdata.py script has undergone significant restructuring with more concise control flows and adjustments for better data handling. This includes the direct addition of transformed 'df_temp' to 'df_list', the shifting of 'concat' method to the end of the script, and the implementation of pandas. Also, a new function named 'plotAdjvsNonAdj' has been added to test_aapl.py, which performs time series plotting for logarithmic and adjusted logarithmic closing prices of selected stock.